requirements.txt
Pyrogram==2.0.79
tgcrypto==1.2.5
python-dotenv==0.21.0